+++ src/fi/tamk/anpro/WeaponEmp.java
@@ -34,7 +34,7 @@ public class WeaponEmp extends AbstractWeapon
      * @param int Ampujan Y-koordinaatti
      */
     @Override
-    public final void activate(int _targetX, int _targetY, float _startX, float _startY)
+    public final void activate(float _targetX, float _targetY, float _startX, float _startY)
     {
         // Tarkistetaan onko ammus jo aktiivinen
         if (!projectile.active) {
